requirements.txt
Django>=4.2.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
hypothesis>=6.82.0